+++ src/routes/(app)/planning/life/+page.svelte
@@ -417,7 +417,10 @@
             future probabilistically
           </p>
         </div>
-        <a class="button is-light" href="/planning/life/goals">Manage Goals</a>
+        <div class="buttons">
+          <a class="button is-light" href="/planning/life/whatif">What-If Scenarios</a>
+          <a class="button is-light" href="/planning/life/goals">Manage Goals</a>
+        </div>
       </div>
     </div>